replace Format type with direct string tokens
There was a place this iota-enum was cast to an int, which depended on
the iota decimal value. The format tokens were changed to strings, and
the precision value derived from them has been changed to use a switch
on the token. This removes the possibility that changing the consts will
break the behaviour of calculation precision.

Token names were also changed and documented.